Reorganize transposonmapping directories
