+++ satay/transposonmapping/exporting/__init__.py
@@ -5,3 +5,11 @@ from .save_per_gene_insertions import (
     save_per_gene_insertions,
 )
 from .export_as_wig import export_as_wig
+
+__all__ = [
+    "export_as_wig",
+    "save_as_bed",
+    "save_per_gene",
+    "save_per_essential_insertions",
+    "save_per_gene_insertions",
+]
