+++ satay/transposonmapping/importing/__init__.py
@@ -2,3 +2,8 @@ from .verify_data_files import verify_data_files
 from .load_default_files import load_default_files
 from .read_genes import read_genes
 from .loading_files import access_files
+
+__all__ = [
+    "load_default_files",
+    "read_genes",
+]
