+++ satay/transposonmapping/mapping/samflag.py
