--- satay/transposonmapping/python_modules/__init__.py
@@ -1,25 +1,0 @@
-from .chromosome_and_gene_positions import gene_position
-from .gene_names import gene_aliases
-from .samflag import samflags
-from .get_chromosome_properties import (
-    get_chromosome_names,
-    get_sequence_length,
-    get_chromosome_reads,
-)
-from .get_reads import get_reads
-from .concatenate_chromosomes import add_chromosome_length
-from .concatenate_chromosomes import add_chromosome_length_inserts
-from .get_insertions_and_reads import get_insertions_and_reads
-
-__all__ = [
-    add_chromosome_length,
-    add_chromosome_length_inserts,
-    gene_aliases,
-    gene_position,
-    get_chromosome_names,
-    get_chromosome_reads,
-    get_insertions_and_reads,
-    get_reads,
-    get_sequence_length,
-    samflags,
-]
