+++ satay/transposonmapping/transposonmapping_satay.py
@@ -3,7 +3,7 @@
 """
 This is a tool developed for analysing transposon insertions for experiments using SAturated Transposon Analysis in Yeast (SATAY).
 This python code contains one function called transposonmapper().
-For more information about this code and the project, see github.com/Gregory94/LaanLab-SATAY-DataAnalysis
+For more information about this code and the project, see https://satay-ll.github.io/SATAY-jupyter-book/Introduction.html
 
 This code is based on the Matlab code created by the Kornmann lab which is available at: sites.google.com/site/satayusers/
 
@@ -22,9 +22,12 @@ Version history:
 import pysam
 
 # Local imports
-from .python_modules import (
+from .properties import (
     get_chromosome_names,
     get_sequence_length,
+)
+
+from .mapping import (
     get_reads,
     add_chromosome_length,
     add_chromosome_length_inserts,
@@ -33,17 +36,17 @@ from .python_modules import (
 
 from .utils import chromosomename_roman_to_arabic
 
+from .importing import (
+    load_default_files,
+    read_genes,
+)
+
 from .exporting import (
     save_as_bed,
     save_per_gene,
     save_per_gene_insertions,
     save_per_essential_insertions,
     export_as_wig,
-)
-
-from .importing import (
-    load_default_files,
-    read_genes,
 )
 
 
@@ -90,6 +93,12 @@ def transposonmapper(bamfile, gff_file=None, essential_file=None, gene_name_file
     for filetype, file_path in data_files.items():
         assert file_path, f"{filetype} not found at {file_path}"
 
+    # Read files for all genes and all essential genes
+    print("Getting coordinates of all genes ...")
+    gene_coordinates, essential_coordinates, aliases_designation = read_genes(
+        gff_file, essential_file, gene_name_file
+    )
+
     # Read bam file
     bam = pysam.AlignmentFile(bamfile, "rb")
 
@@ -106,12 +115,6 @@ def transposonmapper(bamfile, gff_file=None, essential_file=None, gene_name_file
 
     # Get all reads within a specified genomic region
     readnumb_array, tncoordinates_array, tncoordinatescopy_array = get_reads(bam)
-
-    # Read files for all genes and all essential genes
-    print("Getting coordinates of all genes ...")
-    gene_coordinates, essential_coordinates, aliases_designation = read_genes(
-        gff_file, essential_file, gene_name_file
-    )
 
     #%% CONCATENATE ALL CHROMOSOMES
