+++ tests/test_transposonmaper/test_chromosome_position.py
@@ -3,7 +3,7 @@
 import os
 import pkg_resources
 
-from satay.transposonmapping.python_modules.chromosome_and_gene_positions import (
+from satay.transposonmapping.properties.get_chromosome_position import (
     chromosome_position,
 )
 
@@ -15,14 +15,17 @@ def test_packaged_data():
     data_file = "Saccharomyces_cerevisiae.R64-1-1.99.gff3"
     filepath = os.path.join(data_path, data_file)
 
-    chr_length_dict, chr_start_pos_dict, chr_end_pos_dict = chromosome_position(gff_file=filepath)
+    chr_length_dict, chr_start_pos_dict, chr_end_pos_dict = chromosome_position(
+        gff_file=filepath
+    )
 
     assert isinstance(chr_length_dict, dict), "Expected dict type"
     assert isinstance(chr_start_pos_dict, dict), "Expected dict type"
-    assert isinstance(chr_end_pos_dict,dict), "Expected dict type"
+    assert isinstance(chr_end_pos_dict, dict), "Expected dict type"
     assert len(chr_start_pos_dict) == 17, "Expected dictionary of size 17"
     assert len(chr_end_pos_dict) == 17, "Expected dictionary of size 17"
     assert len(chr_length_dict) == 17, "Expected dictionary of size 17"
-    assert  chr_length_dict['II']==813184, "Expected value to be present"
-    assert chr_start_pos_dict['II']==230219 ,"Expected value to be present"
-    assert chr_end_pos_dict[ 'II']==1043402, "Expected value to be present"
+    assert chr_length_dict["II"] == 813184, "Expected value to be present"
+    assert chr_start_pos_dict["II"] == 230219, "Expected value to be present"
+    assert chr_end_pos_dict["II"] == 1043402, "Expected value to be present"
+
